% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fan_util_fun.R
\name{admin_censor_cmprisk}
\alias{admin_censor_cmprisk}
\title{admin_censor_cmprisk}
\usage{
admin_censor_cmprisk(
  df,
  evt_time,
  evt,
  adm_cnr_time = NULL,
  evt_label = NULL,
  overwrite_var = FALSE
)
}
\arguments{
\item{df}{input data}

\item{evt_time}{a numeric vector recording the time points at which the event occurs.}

\item{evt}{a factor vector indicating right censoring (0= censored; 1= event of interest; other= competing risk(s)).}

\item{adm_cnr_time}{a numeric vector specifying the time point at which administrative censoring is applied.}

\item{evt_label}{a numeric vector specifying the time point at which administrative censoring is applied.}

\item{overwrite_var}{a logical scalar (default= FALSE) indiciates if the existing time-to-event variables should be overwritten.}
}
\value{
The input data plus censored time-to-event variables.
}
\description{
admin_censor_cmprisk
}
\details{
The function creates time-to-event variables with the application of administrative censoring for a competing risk
analysis. The newly created variables are named by the same variables names but with a suffix of '_adm' by default.
The original variables can be overwritten by specifying overwrite_var= TRUE. Overwriting the original variables is
not recommended, but it can be useful in some situation.
}
\examples{
cmprisk_df<- read.csv2("http://www.stat.unipg.it/luca/misc/bmt.csv")
admin_censor_cmprisk(cmprisk_df, ftime, status, evt_label = c("0"= "Event free", "1"= "Event", "2"= "Competing event"), adm_cnr_time= 10)
}
