% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fan_util_fun.R
\name{admin_censor_surv}
\alias{admin_censor_surv}
\title{admin_censor_surv}
\usage{
admin_censor_surv(
  df,
  evt_time,
  evt,
  adm_cnr_time = NULL,
  overwrite_var = FALSE
)
}
\arguments{
\item{df}{input data}

\item{evt_time}{a numeric vector recording the time points at which the event occurs.}

\item{evt}{an integer vector indicating right censoring (0= censored; 1= event).}

\item{adm_cnr_time}{a numeric scalar specifying the time point at which administrative censoring is applied.}

\item{overwrite_var}{a logical scalar (default= FALSE) indiciates if the existing time-to-event variables should be overwritten.}
}
\value{
The input data plus censored time-to-event variables.
}
\description{
admin_censor_surv
}
\details{
The function creates time-to-event variables with the application of administrative censoring for a binary (survival) process.
The newly created variables are named by the same variables names but with a suffix of '_adm' by default. The original
variables can be overwritten by specifying overwrite_var= TRUE. Overwriting the original variables is not recommended, but
it can be useful in some situations.
}
\examples{
aml \%>\% admin_censor_surv(evt_time= time, evt= status) # No admin censoring
aml \%>\% admin_censor_surv(evt_time= time, evt= status, adm_cnr_time= 30)
aml \%>\% admin_censor_surv(evt_time= time, evt= status, adm_cnr_time= 30, overwrite_var= TRUE)
}
