% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fan_util_fun.R
\name{decimalplaces}
\alias{decimalplaces}
\title{decimalplaces}
\usage{
decimalplaces(x, max_dec = 4L)
}
\arguments{
\item{x}{a continous variable}

\item{max_dec}{the maximum number of decimal places to return (default 4)}
}
\value{
the most frequent number of digits in the variable
}
\description{
decimalplaces
}
\details{
An internal function that determines the number of digits in the summary of a continuous variable.
}
