% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_desp.R
\name{factor_desp}
\alias{factor_desp}
\title{factor_desp}
\usage{
factor_desp(df, group, includeNA = FALSE)
}
\arguments{
\item{df}{Dataframe}
}
\value{
a dataframe consisting of columns of character variables indicating the frequency and proportion of logical variables.
}
\description{
factor_desp
}
\details{
An internal function that calculate frequencies and proportion of levels in factor/categorical variables.
If there are more than 1 group, then fisher exact tests are applied to assess the between-group differences.
}
