% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanetc-package.R
\docType{package}
\name{fanetc-package}
\alias{fanetc}
\alias{fanetc-package}
\title{fanetc: Descriptive table}
\description{
Miscellaneous functions that Steve wrote to facilitate day-to-day data-analysis projects.
}
\keyword{internal}
