% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_desp.R
\name{fisher_test}
\alias{fisher_test}
\title{fisher_test}
\usage{
fisher_test(df, group)
}
\arguments{
\item{df}{Dataframe}
}
\value{
a dataframe of a single column of character variables indicating p-values.
}
\description{
fisher_test
}
\details{
An internal function that applies fisher exact tests to assess the between-group differences in logical variables
}
