% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_desp.R
\name{k_sample_test}
\alias{k_sample_test}
\title{k_sample_test}
\usage{
k_sample_test(df, group)
}
\arguments{
\item{x}{Numeric variable}

\item{grp}{Factor variable}
}
\value{
a 2-tuple vectors reporting p-values
}
\description{
k_sample_test
}
\details{
An internal function that tests equality of location parameter using one way ANOVA with unequal
variance when mean is reported and nonparametric Kruskal-Wallis tests otherwise in comparing >2 groups.
}
