% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_desp.R
\name{logical_desp}
\alias{logical_desp}
\title{logical_desp}
\usage{
logical_desp(df, group)
}
\arguments{
\item{df}{Dataframe}
}
\value{
a dataframe consisting of columns of character variables indicating the frequency and proportion of logical variables.
}
\description{
logical_desp
}
\details{
An internal function that calculate frequencies and proportion of TRUE in logical variables. If there are more
than 1 group, then fisher exact tests are applied to assess the between-group differences.
}
