% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desp_table_gtsummary.R
\name{mean_sd}
\alias{mean_sd}
\title{Helper functions for custom statistics}
\usage{
mean_sd(x)
}
\arguments{
\item{x}{Numeric variable}
}
\value{
Character string with formatted mean +/- SD
}
\description{
These functions work with gtsummary's inline_text() and other functions to extract
specific statistics from the summary table.

Format mean and standard deviation
}
