% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desp_table_gtsummary.R
\name{med_iqr}
\alias{med_iqr}
\title{med_iqr}
\usage{
med_iqr(x)
}
\arguments{
\item{x}{Numeric variable}
}
\value{
Character string with formatted median (Q1-Q3)
}
\description{
Format median and interquartile range
}
