% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desp_table_gtsummary.R
\name{n_avail}
\alias{n_avail}
\title{n_avail}
\usage{
n_avail(x)
}
\arguments{
\item{x}{Variable}
}
\value{
Formatted count of non-missing values
}
\description{
Count available (non-missing) observations
}
