% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fan_util_fun.R
\name{recode_missing}
\alias{recode_missing}
\title{recode_missing}
\usage{
recode_missing(x, na.value = NULL)
}
\value{
input variable with NA
}
\description{
recode_missing
}
\details{
An internal function that replace missing value code with NA.
}
