% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fan_util_fun.R
\name{calculate_type3_mi}
\alias{calculate_type3_mi}
\title{calculate_type3_mi}
\usage{
calculate_type3_mi(mira_obj, vcov_fun = NULL)
}
\description{
calculate_type3_mi
}
\details{
The function calculates the  3 p-values based on Wald's statistics.
}
