% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fan_util_fun.R
\name{construct_cmprisk_var}
\alias{construct_cmprisk_var}
\title{construct_cmprisk_var}
\usage{
construct_cmprisk_var(
  df,
  patid,
  idx_dt,
  evt_dt,
  end_dt,
  cmprisk_varname = NULL,
  append = FALSE,
  ...
)
}
\description{
construct_cmprisk_var
}
\details{
The function creates time-to-event variables for competing risk data
}
