% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_time_desp.R
\name{estimate_cif}
\alias{estimate_cif}
\title{estimate_cif}
\usage{
estimate_cif(df, evt_time, evt, group, ...)
}
\description{
estimate_cif
}
\details{
The function analyzes the competing data (df) using Andersen-Johansen method in estimating cumulative incidence
function.The function store the input data in the call(), which can be used in run_gray_test().
}
