% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_time_desp.R
\name{estimate_km}
\alias{estimate_km}
\title{estimate_km}
\usage{
estimate_km(df, evt_time, evt, group, ...)
}
\description{
estimate_km
}
\details{
The function analyzes the data (df) using Kaplan-Meier survival method with pointwise 95% CI estimated using log-log
transformation (same as SAS's defualt). The function store the input data in the call(), which can be used in
run_logrank_test().
}
