% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fan_util_fun.R
\name{format_pvalue}
\alias{format_pvalue}
\title{format.pvalue}
\usage{
format_pvalue(x, eps = 0.001, trim = TRUE, droptrailing0 = FALSE,
  pad = FALSE, ...)
}
\arguments{
\item{x}{Numeric variable}
}
\value{
character variables reporting p-values
}
\details{
An internal function that formats p-values according to the statistical guidelines of the Annals of Medicine.
}
