% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_desp.R
\name{mean_sd}
\alias{mean_sd}
\title{mean_sd}
\usage{
mean_sd(x)
}
\arguments{
\item{x}{Numeric variable}
}
\value{
a character reporting the mean plus/minus standard deviation
}
\description{
mean_sd
}
\details{
An internal function that format mean and standard deviation of a continuous variable
}
