% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_desp.R
\name{med_iqr}
\alias{med_iqr}
\title{med_iqr}
\usage{
med_iqr(x)
}
\arguments{
\item{x}{Numeric variable}
}
\value{
a character reporting the median (Q1 - Q3)
}
\description{
med_iqr
}
\details{
An internal function that format median and interquartile range of a continuous variable
}
