% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_desp.R
\name{n_avail}
\alias{n_avail}
\title{n_avail}
\usage{
n_avail(x)
}
\arguments{
\item{x}{Numeric variable}
}
\value{
a character reporting the number of non-missing observations
}
\description{
n_avail
}
\details{
An internal function that format the number of non-missing observations in a variable
}
