% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_time_desp.R
\name{show_cif}
\alias{show_cif}
\title{show_cif}
\usage{
show_cif(surv_obj, evt_type = 1, evt_label = function(x) {    
  recode_factor(x, `1` = "Event", `2` = "Competing event", .default =
  "Event free") }, add_ci = TRUE, add_atrisk = TRUE,
  add_legend = FALSE, add_pvalue = TRUE,
  pvalue_pos = c("bottomright", "topleft", "topright", "bottomleft",
  "left", "right", "top", "bottom"), plot_theme = theme_minimal(),
  x_lab = "Time", y_lab = "Proportion of subjects", x_lim = NULL,
  y_lim = NULL, x_break = NULL, y_break = NULL,
  color_scheme = c("brewer", "grey", "viridis", "manual"),
  color_list = NULL, plot_cdf = FALSE)
}
\arguments{
\item{surv_obj}{a survfitms subject.}

\item{add_ci}{an integer vector indicating right censoring (0= censored; 1= event of interest; other= competing risk(s)).}

\item{add_atrisk}{a numeric vector specifying the time point at which administrative censoring is applied.}

\item{add_legend}{a numeric vector specifying the time point at which administrative censoring is applied.}

\item{add_pvalue}{a logical scalar (default= FALSE) indiciates if the existing time-to-event variables should be overwritten.}

\item{pvalue_pos}{a logical scalar (default= FALSE) indiciates if the existing time-to-event variables should be overwritten.}

\item{plot_theme}{a numeric vector recording the time points at which the event occurs.}

\item{x_lab}{an integer vector indicating right censoring (0= censored; 1= event of interest; other= competing risk(s)).}

\item{y_lab}{a numeric vector specifying the time point at which administrative censoring is applied.}

\item{x_lim}{a numeric vector specifying the time point at which administrative censoring is applied.}

\item{y_lim}{a logical scalar (default= FALSE) indiciates if the existing time-to-event variables should be overwritten.}

\item{color_list}{input data}

\item{plot_cdf}{a logical scalar (default= FALSE) indiciates if the existing time-to-event variables should be overwritten.}
}
\value{
A ggplot object.
}
\description{
show_cif
}
\details{
The function shows the cumulative incidence function for competing risks with and without strata.
}
