% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_time_desp.R
\name{show_surv}
\alias{show_surv}
\title{show_surv}
\usage{
show_surv(
  surv_obj,
  x_lab = "Time",
  y_lab = if (plot_cdf) "The proportion of deceased subjects" else
    "The freedom from death",
  y_lim = NULL,
  x_break = NULL,
  y_break = NULL,
  color_scheme = c("brewer", "grey", "viridis", "manual"),
  color_list = NULL,
  plot_theme = theme_minimal(),
  add_ci = TRUE,
  add_atrisk = TRUE,
  add_legend = FALSE,
  add_pvalue = TRUE,
  atrisk_init_pos = NULL,
  pvalue_pos = c("topleft", "topright", "bottomleft", "bottomright", "left", "right",
    "top", "bottom"),
  plot_cdf = FALSE
)
}
\arguments{
\item{surv_obj}{a numeric vector recording the time points at which the event occurs.}

\item{x_lab}{an integer vector indicating right censoring (0= censored; 1= event of interest; other= competing risk(s)).}

\item{y_lab}{a numeric vector specifying the time point at which administrative censoring is applied.}

\item{y_lim}{a logical scalar (default= FALSE) indiciates if the existing time-to-event variables should be overwritten.}

\item{color_list}{a list containing the parameter values for scale_color_manual() or scale_fill_manual()}

\item{plot_theme}{a user-specify ggplot theme.}

\item{add_ci}{logical parameter indicating whether a pointwise 95% confidence interval should be added.}

\item{add_atrisk}{a logical parameter indicating whether at-risk table should be added to the figure.}

\item{add_legend}{a logical parameter indicating whether legend should be added to the figure.}

\item{add_pvalue}{a logical parameter (default= FALSE) indiciates if a p-value should be added to the plot.}

\item{pvalue_pos}{a character parameter indicating where the p-value should be added to the plot.}

\item{plot_cdf}{a logical scalar (default= FALSE) indiciates if the CDF (instead of) survival function should be plotted.}

\item{x_lim}{a numeric vector specifying the time point at which administrative censoring is applied.}
}
\value{
A ggplot object.
}
\description{
show_surv
}
\details{
The function shows the survival function with and without strata.
}
