% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fan_util_fun.R
\name{summarize_cif}
\alias{summarize_cif}
\title{summarize_cif}
\usage{
summarize_cif(fit, times = NULL)
}
\description{
summarize_cif
}
\details{
The function summarizes the fitted CIF at the time points specified by a user.
}
