% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fan_util_fun.R
\name{summarize_coxph}
\alias{summarize_coxph}
\title{summarize_coxph}
\usage{
summarize_coxph(mdl, exponentiate = TRUE, maxlabel = 100, alpha = 0.05)
}
\description{
summarize_coxph
}
\details{
The function summarizes the fitted cox model with the type 3 error based on Wald's statistics.
}
