% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fan_util_fun.R
\name{summarize_km}
\alias{summarize_km}
\title{summarize_km}
\usage{
summarize_km(fit, times = NULL, failure_fun = FALSE)
}
\description{
summarize_km
}
\details{
The function summarize the fitted KM at the time points specified by a user.
}
